添加序列化器，动态加载配置，利用SPI机制，加载自定义的序列化器实现类

+++ common/src/main/java/com/samoyer/common/service/UserService.java
@@ -8,15 +8,18 @@ import com.samoyer.common.model.po.User;
 public interface UserService {
     /**
      * 获取用户
+     * @param user
+     * @return
      */
     User getUserInfo(User user);
 
     /**
      * 获取数字
      * mock
+     * default表示允许在接口中定义有具体实现的方法，不要求实现类必须重写
+     * @return
      */
-    //default表示允许在接口中定义有具体实现的方法，不要求实现类必须重写
-    default short getNumber(){
+    default Integer getNumber(){
         return 1;
     }
 }
